Refactor Capabilities
Eliminates the distinction of "Capability" versus "Capabilities" and
treats the latter as bitflags.
